requirements.txt
# CryoSure Sensor Simulation Dependencies

# AWS IoT Python SDK for MQTT communication
AWSIoTPythonSDK==1.4.12

# Python dotenv for environment variable management
python-dotenv==1.0.0
